updated source code

+++ back_end/fruitsapp_backend/Models/PriceProduct.cs
@@ -8,7 +8,7 @@ namespace fruitsapp_backend.Models
         [Key]
         public int Id { set; get; }
         public int product_id { set; get; }
-        public int? price { set; get; }
+        public decimal? price { set; get; }
         public string? currency { set; get; }
         public string? description { set; get; }
         public DateTime create_at { set; get; }
